Add a PushSubscriptionOptions dictionary with a userVisible option.
User agents may decide to display different UI when requesting push
permission from the user if the intended use of the subscription is to
only send messages which result in user visible UI, for example a
Web Notification.

This PR introduces this in the specification by accepting an optional
property bag to the subscribe() and hasPermission() methods. This bag
features a "userVisible" option that can be set to TRUE.

UAs are not required to follow this UI paradigm, but when they do, they
SHOULD enforce the requirement when receiving a push message. The exact
tools for "user visible UI" and enforcement are left undefined as
one can imagine multiple solutions for them, which we don't want to
capture in the specification.