Add PushSubscription.expirationTime
The expiration time of a push subscription indicates the time, in
milliseconds since the UNIX epoch (UTC), at which the subscription will
be deactivated.

This is an important addition for user agents that refresh their
subscriptions, as it gives the developer knowledge of a final time after
which they can garbage collect the stored subscriptions if the user has
not communicated the details of the refreshed subscription.

Fixes #86