Add Declarative Web Push
This introduces a new feature whereby push messages conforming to a certain JSON format directly create an end user notification and show it (possibly preceded by a new pushnotification event).

In addition to showing a notification, the app badge can be updated as well.

This builds on https://github.com/whatwg/notifications/pull/213 which adds URL members to notifications.

Exposing PushManager outside of service workers will be done in a separate change.